+++ server/apps/server-app/build.gradle.kts
@@ -103,7 +103,7 @@ dependencies {
     implementation(project(":server:libs:platform:platform-user:platform-user-rest:platform-user-rest-impl"))
     implementation(project(":server:libs:platform:platform-user:platform-user-service"))
     implementation(project(":server:libs:platform:platform-webhook:platform-webhook-impl"))
-    implementation(project(":server:libs:platform:platform-webhook:platform-webhook-rest:platform-webhook-rest-impl"))
+    implementation(project(":server:libs:platform:platform-webhook:platform-webhook-web:platform-webhook-web-impl"))
     implementation(project(":server:libs:platform:platform-workflow:platform-workflow-coordinator:platform-workflow-coordinator-impl"))
     implementation(project(":server:libs:platform:platform-workflow:platform-workflow-task-dispatcher:platform-workflow-task-dispatcher-service"))
     implementation(project(":server:libs:platform:platform-workflow:platform-workflow-test:platform-workflow-test-rest"))
