+++ settings.gradle.kts
@@ -180,8 +180,8 @@ include("server:libs:platform:platform-user:platform-user-rest:platform-user-res
 include("server:libs:platform:platform-user:platform-user-service")
 include("server:libs:platform:platform-webhook:platform-webhook-api")
 include("server:libs:platform:platform-webhook:platform-webhook-impl")
-include("server:libs:platform:platform-webhook:platform-webhook-rest:platform-webhook-rest-api")
-include("server:libs:platform:platform-webhook:platform-webhook-rest:platform-webhook-rest-impl")
+include("server:libs:platform:platform-webhook:platform-webhook-web:platform-webhook-web-api")
+include("server:libs:platform:platform-webhook:platform-webhook-web:platform-webhook-web-impl")
 include("server:libs:platform:platform-workflow:platform-workflow-coordinator:platform-workflow-coordinator-api")
 include("server:libs:platform:platform-workflow:platform-workflow-coordinator:platform-workflow-coordinator-impl")
 include("server:libs:platform:platform-workflow:platform-workflow-task-dispatcher:platform-workflow-task-dispatcher-api")
